DBConnect Class Change
DBConnectClass.cs added to keep program.cs clean
