+++ DataEntryTestApp/DataEntryTestApp/DataEntryTestApp.csproj
@@ -52,6 +52,7 @@
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="ClientObjects\Event.cs" />
+    <Compile Include="DBConnect.cs" />
     <Compile Include="Forms\Creation\EmployeeInitializer.cs">
       <SubType>Form</SubType>
     </Compile>
