+++ wp
@@ -1,0 +1,7 @@
+Merge branch 'Sliferss-master'
+# Please enter a commit message to explain why this merge is necessary,
+# especially if it merges an updated upstream into a topic branch.
+#
+# Lines starting with '#' will be ignored, and an empty message aborts
+# the commitPost Form Added for communication with server:wq
+the commit post form added for communication wtih server
